int->size_t for-loop i constructBoardFromSolution in solutions.cpp

+++ src/solutions.cpp
@@ -8,7 +8,7 @@ Board Solutions::constructBoardFromSolution(const Solution& solution, const std:
     for (int row = 0; row < 10; ++row)
         for (int col = 0; col < 10; ++col)
             board[10 * row + col] = ' ';
-    for (int i = 0; i < pieceList.size(); ++i)
+    for (size_t i = 0; i < pieceList.size(); ++i)
         slvr::Solver::placePiece(pcs::piece_names.at(pieceList[i]), solution[i].second, solution[i].first, &board);
     return board;
 }
